# Conflicts:
#	ForestGame/.vs/ForestGame/v14/.suo
#	ForestGame/Library/CurrentLayout.dwlt
#	ForestGame/Library/InspectorExpandedItems.asset
#	ForestGame/Library/ScriptAssemblies/Assembly-CSharp.dll
#	ForestGame/Library/ScriptAssemblies/Assembly-CSharp.dll.mdb
#	ForestGame/Library/ScriptAssemblies/Assembly-UnityScript.dll
#	ForestGame/Library/ScriptAssemblies/Assembly-UnityScript.dll.mdb
#	ForestGame/Library/assetDatabase3
#	ForestGame/Library/expandedItems
#	ForestGame/Library/metadata/00/00000000000000002000000000000000
#	ForestGame/Library/metadata/00/00000000000000003000000000000000
#	ForestGame/Library/metadata/00/00000000000000004000000000000000
#	ForestGame/Library/metadata/00/00000000000000004100000000000000
#	ForestGame/Library/metadata/00/00000000000000005000000000000000
#	ForestGame/Library/metadata/00/00000000000000005100000000000000
#	ForestGame/Library/metadata/00/00000000000000006000000000000000
#	ForestGame/Library/metadata/00/00000000000000006100000000000000
#	ForestGame/Library/metadata/00/00000000000000007000000000000000
#	ForestGame/Library/metadata/00/00000000000000007100000000000000
#	ForestGame/Library/metadata/00/00000000000000008000000000000000
#	ForestGame/Library/metadata/00/00000000000000009000000000000000
#	ForestGame/Library/metadata/00/0000000000000000a000000000000000
#	ForestGame/Library/metadata/00/0000000000000000a100000000000000
#	ForestGame/Library/metadata/00/0000000000000000c000000000000000
#	ForestGame/Library/metadata/90/90943f3da4a49f34db97b78ed2e26ce7.info

+++ ForestGame/Assets/UnityVS/Influences.cs
@@ -1,8 +1,8 @@
 ﻿public class Influences {
 
-    private byte industrialInfluence;
-    private byte environmentalInfluence;
-    private byte touristicInfluence;
+    private short industrialInfluence;
+    private short environmentalInfluence;
+    private short touristicInfluence;
 
     public Influences() {
         industrialInfluence = 0;
@@ -10,27 +10,27 @@
         touristicInfluence = 0;
     }
 
-    public void setIndustrialInfluence(byte influence) {
+    public void setIndustrialInfluence(short influence) {
         industrialInfluence = influence;
     }
 
-    public byte getIndustrialInfluence() {
+    public short getIndustrialInfluence() {
         return industrialInfluence;
     }
 
-    public void setEnvironmentalInfluence(byte influence) {
+    public void setEnvironmentalInfluence(short influence) {
         industrialInfluence = influence;
     }
 
-    public byte getEnvironmentalInfluence() {
+    public short getEnvironmentalInfluence() {
         return environmentalInfluence;
     }
 
-    public void setTouristicalInfluence(byte influence) {
+    public void setTouristicalInfluence(short influence) {
         touristicInfluence = influence;
     }
 
-    public byte getTouristicalInfluence() {
+    public short getTouristicalInfluence() {
         return touristicInfluence;
     }
